requirements.txt
asgiref==3.7.2
Django==4.2.8
sqlparse==0.4.4
typing_extensions==4.8.0
djangorestframework==3.14.0
pytz==2023.3.post1
django-simple-captcha==0.6.0
mysqlclient